Move example code in main.rs to a proper example
